+++ examples/deleteme.js
@@ -1,12 +1,12 @@
 // node --experimental-modules enumerate-segments.js
-import {ExifParser} from '../src/index-full.js'
+import Exifr from '../src/index-full.js'
 import {promises as fs} from 'fs'
 import path from 'path'
 
 (async function() {
 	let filePath = path.join('../test/fixtures/001.tif')
 	let fileBuffer = await fs.readFile(filePath)
-	let exifr = new ExifParser({wholeFile: true, mergeOutput: false})
+	let exifr = new Exifr({wholeFile: true, mergeOutput: false})
 	await exifr.read(fileBuffer)
 	exifr.parse()
 })()