+++ examples/enumerate-segments.js
@@ -1,5 +1,5 @@
 // node --experimental-modules enumerate-segments.js
-import {ExifParser} from '../src/index-full.js'
+import Exifr from '../src/index-full.js'
 import {promises as fs} from 'fs'
 import path from 'path'
 
@@ -9,7 +9,7 @@ import path from 'path'
 	for (let fileName of imageFiles) {
 		let filePath = path.join('../test/fixtures/', fileName)
 		let fileBuffer = await fs.readFile(filePath)
-		let exifr = new ExifParser({wholeFile: true, mergeOutput: false})
+		let exifr = new Exifr({wholeFile: true, mergeOutput: false})
 		await exifr.read(fileBuffer)
 		exifr.parse()
 		console.log('----------------------------------------------------')