+++ src/index-full.js
@@ -1,4 +1,3 @@
-export * from './index-lite.js'
 import Exifr from './index-lite.js'
 export default Exifr